move fsync/fdatasync, add linkat, symlinkat
